+++ lib/generators/logidze/install/functions/logidze_version.sql
@@ -1,4 +1,4 @@
-CREATE OR REPLACE FUNCTION logidze_version(v bigint, data jsonb, ts timestamp with time zone, blacklist text[] DEFAULT '{}') RETURNS jsonb AS $body$
+CREATE OR REPLACE FUNCTION logidze_version(v bigint, data jsonb, ts timestamp with time zone) RETURNS jsonb AS $body$
   DECLARE
     buf jsonb;
   BEGIN
@@ -8,7 +8,7 @@ CREATE OR REPLACE FUNCTION logidze_version(v bigint, data jsonb, ts timestamp wi
               'v',
               v,
               'c',
-              logidze_exclude_keys(data, VARIADIC array_append(blacklist, 'log_data'))
+              data
               );
     IF coalesce(current_setting('logidze.meta', true), '') <> '' THEN
       buf := jsonb_insert(buf, '{m}', current_setting('logidze.meta')::jsonb);