+++ spec/acceptance_helper.rb
@@ -4,6 +4,7 @@ require "spec_helper"
 
 RSpec.configure do |config|
   config.include Logidze::AcceptanceHelpers
+  config.include Logidze::SqlHelpers
 
   config.around(:each) do |example|
     Dir.chdir("#{File.dirname(__FILE__)}/dummy") do